+++ Cryptographic_Backend/src/cb/backend/KeyStorage.java
@@ -5,8 +5,8 @@ import java.security.PublicKey;
 
 /* La idea es que estas interfaces sean implementadas como un Singleton :) */
 public interface KeyStorage {
-	abstract public int storeKeyPair(String keyType, int keySize, KeyPair pair);
-	abstract public PrivateKey getPrivateKey(int handler);
-	abstract public PublicKey getPublicKey(int handler);
-	abstract public String getAttribute(String attr, int handler) throws Exception;
+	abstract public long storeKeyPair(String keyType, int keySize, KeyPair pair);
+	abstract public PrivateKey getPrivateKey(long handler);
+	abstract public PublicKey getPublicKey(long handler);
+	abstract public String getAttribute(String attr, long handler) throws Exception;
 }
